refactor: JavascriptUtil 리팩토링
